+++ src/main/kotlin/com/arnyminerz/database/dsl/Events.kt
@@ -19,10 +19,10 @@ object Events : LongIdTable(), JsonSerializer<Event?> {
     val date: Column<String> = varchar("date", DSLConst.DATE_LENGTH)
     val until: Column<String?> = varchar("until", DSLConst.DATE_LENGTH).nullable()
     val reservations: Column<String?> = varchar("reservations", DSLConst.DATE_LENGTH).nullable()
-    val maxGuests: Column<Int?> = integer("max_guests").nullable().default(-1)
+    val maxGuests: Column<Long?> = long("max_guests").nullable().default(-1)
     val publicKey: Column<String> = varchar("public_key", EVENT_KEY_LENGTH)
     val privateKey: Column<String> = varchar("private_key", EVENT_KEY_LENGTH)
 
     override suspend fun fromJson(json: JSONObject): Event? =
-        ServerDatabase.instance.eventsInterface.get(json.getInt("id")) { it }
+        ServerDatabase.instance.eventsInterface.get(json.getLong("id")) { it }
 }
