+++ src/main/kotlin/com/arnyminerz/database/dsl/Transactions.kt
@@ -10,7 +10,7 @@ object Transactions : LongIdTable() {
     val timestamp: Column<String> = varchar("timestamp", DATE_LENGTH)
     val date: Column<String> = varchar("date", DATE_LENGTH)
     val pricePerUnit: Column<Double> = double("unit_price")
-    val amount: Column<Int> = integer("amount")
+    val amount: Column<Long> = long("amount")
     val description: Column<String> = varchar("description", TRANSACTION_DESCRIPTION_LENGTH)
 
     val user = reference("user", Users)
