+++ src/main/kotlin/com/arnyminerz/endpoints/profile/UpdateUserCategoryEndpoint.kt
@@ -16,7 +16,7 @@ import io.ktor.util.pipeline.PipelineContext
 
 object UpdateUserCategoryEndpoint : AuthenticatedEndpoint(Permissions.Users.ChangeCategory) {
     override suspend fun PipelineContext<*, ApplicationCall>.endpoint(user: User) {
-        val userId: Int by call.parameters
+        val userId: Long by call.parameters
 
         val categoryName by called { Arguments.Category }
